refactor: 에러 메세지 폴더명 및 상수명 변경
