+++ client/src/features/Event/New/utils/validateEventForm.ts
@@ -1,5 +1,5 @@
 import { EventFormData } from '../../types/Event';
-import { MAX_LENGTH, VALIDATION_MESSAGES } from '../constants/validation';
+import { MAX_LENGTH, ERROR_MESSAGES } from '../constants/errorMessages';
 
 const isEmpty = (value: string) => value.trim() === '';
 const isTooLong = (value: string) => value.length > MAX_LENGTH;
@@ -15,31 +15,30 @@ export const getValidationMessage = (
 ): string => {
   switch (field) {
     case 'title':
-      if (isEmpty(value)) return VALIDATION_MESSAGES.REQUIRED('이벤트 이름');
-      if (isTooLong(value)) return VALIDATION_MESSAGES.MAX_LENGTH('이벤트 이름', MAX_LENGTH);
+      if (isEmpty(value)) return ERROR_MESSAGES.REQUIRED('이벤트 이름');
+      if (isTooLong(value)) return ERROR_MESSAGES.MAX_LENGTH('이벤트 이름', MAX_LENGTH);
       break;
     case 'eventStart':
-      if (!isFutureDate(value)) return VALIDATION_MESSAGES.EVENT_START_MUST_BE_FUTURE;
+      if (!isFutureDate(value)) return ERROR_MESSAGES.EVENT_START_MUST_BE_FUTURE;
       break;
     case 'eventEnd':
-      if (isBefore(value, formData.eventStart))
-        return VALIDATION_MESSAGES.EVENT_END_MUST_BE_AFTER_START;
+      if (isBefore(value, formData.eventStart)) return ERROR_MESSAGES.EVENT_END_MUST_BE_AFTER_START;
       break;
     case 'registrationEnd':
-      if (!isFutureDate(value)) return VALIDATION_MESSAGES.EVENT_START_MUST_BE_FUTURE;
+      if (!isFutureDate(value)) return ERROR_MESSAGES.EVENT_START_MUST_BE_FUTURE;
       if (isAfterorEqual(value, formData.eventStart))
-        return VALIDATION_MESSAGES.REGISTRATION_DEADLINE_BEFORE_EVENT_START;
+        return ERROR_MESSAGES.REGISTRATION_DEADLINE_BEFORE_EVENT_START;
       break;
     case 'place':
-      if (isEmpty(value)) return VALIDATION_MESSAGES.REQUIRED('장소');
-      if (isTooLong(value)) return VALIDATION_MESSAGES.MAX_LENGTH('장소', MAX_LENGTH);
+      if (isEmpty(value)) return ERROR_MESSAGES.REQUIRED('장소');
+      if (isTooLong(value)) return ERROR_MESSAGES.MAX_LENGTH('장소', MAX_LENGTH);
       break;
     case 'description':
-      if (isEmpty(value)) return VALIDATION_MESSAGES.REQUIRED('설명');
-      if (isTooLong(value)) return VALIDATION_MESSAGES.MAX_LENGTH('설명', MAX_LENGTH);
+      if (isEmpty(value)) return ERROR_MESSAGES.REQUIRED('설명');
+      if (isTooLong(value)) return ERROR_MESSAGES.MAX_LENGTH('설명', MAX_LENGTH);
       break;
     case 'maxCapacity':
-      if (!isPositiveInteger(value)) return VALIDATION_MESSAGES.NOT_POSITIVE_INTEGER;
+      if (!isPositiveInteger(value)) return ERROR_MESSAGES.NOT_POSITIVE_INTEGER;
       break;
   }
